refactor: create useScroll custom hook
scroll관련 효과를 위해 사용하던 ref를 하나로 묶어 useScroll custom hook으로 재정의.
