+++ src/pages/Start/Main/Main.scss
@@ -2,5 +2,5 @@ main.Main {
   @include prop-flex;
   @include selector-scrollbar;
 
-  grid-area: 1 / 2 / 3 / 3;
+  grid-area: 1 / 2 / 4 / 3;
 }
